chore: rename tr-macros.h to macros.h
