+++ gtk/Flags.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include <libtransmission/tr-macros.h>
+#include <libtransmission/macros.h>
 
 #include <bitset>
 #include <initializer_list>
