+++ gtk/TorrentSorter.cc
@@ -12,7 +12,7 @@
 #include <libtransmission-app/display-modes.h>
 
 #include <libtransmission/transmission.h>
-#include <libtransmission/tr-macros.h>
+#include <libtransmission/macros.h>
 #include <libtransmission/utils.h>
 
 #include <small/map.hpp>
