+++ libtransmission/macros.h
