+++ libtransmission/torrent-files.h
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include "libtransmission/file.h"
-#include "libtransmission/tr-macros.h"
+#include "libtransmission/macros.h"
 #include "libtransmission/tr-strbuf.h"
 #include "libtransmission/types.h"
