+++ tests/libtransmission/crypto-test.cc
@@ -17,7 +17,7 @@
 
 #include <libtransmission/peer-mse.h>
 #include <libtransmission/crypto-utils.h>
-#include <libtransmission/tr-macros.h>
+#include <libtransmission/macros.h>
 #include <libtransmission/string-utils.h>
 
 using namespace std::literals;
