+++ tests/libtransmission/file-test.cc
@@ -30,7 +30,7 @@
 #include <libtransmission/error.h>
 #include <libtransmission/file.h>
 #include <libtransmission/string-utils.h>
-#include <libtransmission/tr-macros.h>
+#include <libtransmission/macros.h>
 #include <libtransmission/tr-strbuf.h>
 
 #include "test-fixtures.h"
