+++ tests/libtransmission/magnet-metainfo-test.cc
@@ -9,7 +9,7 @@
 
 #include <libtransmission/crypto-utils.h> // tr_rand_buffer()
 #include <libtransmission/magnet-metainfo.h>
-#include <libtransmission/tr-macros.h>
+#include <libtransmission/macros.h>
 
 #include "test-fixtures.h"
